Fixed Free Cam mode
There was a long standing issue (it's been there since the beginning of
the project) where in Free Cam mode, the camera was still rotating when
you hold fully left or right.

I had fixed this in ArchetypePrototype (AI branch), but it was some
crazy ass way of reading the individual animation rotation velocities,
which is super unmaintainable.